AltAI/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
PyJWT>=2.8.0
huggingface_hub>=0.23.0
requests>=2.31.0

